CP-54351 move metadata from /opt
